--- src/main/java/calculator/CalculatorClient.java
@@ -1,10 +1,0 @@
-package calculator;
-
-import calculator.domain.StringCalculator;
-
-public class CalculatorClient {
-    public static void main(String[] args) {
-        StringCalculator stringCalculator = new StringCalculator();
-        stringCalculator.calculate();
-    }
-}
