+++ src/main/java/calculator/StringCalculator.java
@@ -1,10 +1,11 @@
-package calculator.domain;
+package calculator;
 
+import calculator.domain.Expression;
 import calculator.view.InputView;
 import calculator.view.ResultView;
 
 public class StringCalculator {
-    public void calculate() {
+    public static void main(String[] args) {
         Expression expression = new Expression(InputView.askExpression());
         ResultView.printResult(expression.result());
     }
